Join Room
The user is now able to join rooms, without any complications, their user ID is inserted within the array, the model has also been updated which now allows tasks to be added, it shows when and by who the room was created. The next stage of the code, it will allow the user to view each individual room data, such as the members and tasks.